python-demo/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=6.30.0,<7.0.0
PyYAML>=6.0
base58>=2.1.1
